Generalize `derive(Zeroable)` by adding bounds for generic types to impl
Before, the struct itself already needed to have `T: Zeroable` bounds
in order for the derive to work. However, we can also add those bounds
to the `Zeroable` impl itself. Meaning that the inner types only need
to be zeroable in order to implement `Zeroable` for the outer type.

This is in line with how derives like `Clone` work.