Move `field_ty: Pod` bounds to impl `where` bounds for `derive(Pod)`
Before, the check that all struct field types implement `Pod` was done
in some unnamed `const`. With this commit, those checks are `where`
bounds on the generated impl instead. This is mainly a preparation for
allowing generic parameters for `derive(Pod)` in a few situations. But
it is also slightly simpler than the previous version.