Move the array impl for `Pod` to superset types
This change removes the current `impl Pod for [T; N]` and moves it to
superset types, namely into two other impls: `impl NonUninit for [T; N]`
and `impl AnyBitPattern for [T; N]`.

This, makes it possible for the user to cast between an array of
`NonUnint` and an array of `AnyBitPattern` for example, which was
previously impossible. One example usecase would be to cast between
`[char; N]` and `[u8; N]`.