allow unexpected_cfgs lint
AFAICT it's not possible to set `#[allow(unexpected_cfgs)]` on a
`#[cfg(...)]` or `#[cfg_attr(...)]` directly. Instead, we have to crate
a new scope and use `#[allow(...)]` on that. This patch wraps all uses
of `target_arch = "spirv"` in an unnamed constant scope and uses
`#[allow(unexpected_cfgs)]` on it.