hide `unexpected_cfgs` warnings in proc-macro generated code (#287)
* replace #[derive(Debug)] with manual impls

The new impls should be identical, but they have the advantage that we
can move them around. See the next patch for why this is useful.

* allow unexpected_cfgs lint

AFAICT it's not possible to set `#[allow(unexpected_cfgs)]` on a
`#[cfg(...)]` or `#[cfg_attr(...)]` directly. Instead, we have to crate
a new scope and use `#[allow(...)]` on that. This patch wraps all uses
of `target_arch = "spirv"` in an unnamed constant scope and uses
`#[allow(unexpected_cfgs)]` on it.