add support for deriving NoUninit on enums with fields (#292)
* add support for deriving NoUninit on enums with fields

We check for padding like we do for structs except that we also
consider the enum discriminant when calculating the unpadded size.

* improve support for #[repr(C)] enums

Unfortunately, the integer discriminant used for #[repr(C)] doesn't
have a name (it's not always core::ffi::c_int), but we can use some
compile-time tricks to get the integer type. Use that instead of hard-
coding core::ffi::c_int and add support for deriving NoUninit
for #[repr(C)] enums.

* simplify get_enum_discriminant for enums w/o fields

* add more comments

Co-authored-by: Gray Olson <gray@grayolson.com>

* update requirements for NoUninit

* link type layout section in NoUninit docs

* small wording change on no uninit docs

* inline type alias

This makes the generated code a bit harder to read, but also has the
advantage of not unnecessarily adding a type to the global namespace
for CheckedBitPattern.

---------

Co-authored-by: Gray Olson <gray@grayolson.com>